clean up unused vars/files
