+++ app/src/test-helpers/survey-helpers.ts
@@ -10,7 +10,6 @@ import { geoJsonFeature } from './spatial-helpers';
 
 export const surveyObject: SurveyViewObject = {
   survey_details: {
-    uuid: 'uuid',
     id: 1,
     project_id: 1,
     survey_name: 'survey name',
